Lecture 1: From the one-half rule to the bunching method
There are 3 parts in this exercise. The first part requires estimation of pi using a direct-sampling algorithm, then determines the statistical error of the numerical calculation. In the second part, the estimation method is changed to a Markov-chain algorithm. In the third part, the bunching algorithm is applied to quantify the error associated with the Markov-chain algorithm